+++ CHANGELOG.md
@@ -18,7 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - docbook: removed ToHtml and direct (non-XSLT) processing functionality and Playwright
 - docbook: moved docbook, fop, math and pdf packages from the core module into the docbook module
 - html: moved Html and A into the xml package, and merged the rest of the package into the site package in the collector module
-
+- xml: use XML includes instead of redirects
+- xml: worked around a terrible Xerces XInclude bug
+- xml: removed my home-grown redirect support
+- 
 ## [0.10.7] - 2023-01-28
 - fix: set the fontSize attribute even if its value is the default
 - fix: escape backslashes in TeX strings